refactor: move plain-date helpers into ui and derive today from the app timezone
addDays/daysBetween/isoWeek now live in @lattice-php/ui/format/plain-date as
generic Y-m-d arithmetic; the timeline derives its today marker from
currentTimezone() instead of UTC, which fixed the marker sitting on the wrong
day around midnight for non-UTC users. Axis building and lane packing stay in
the calendar package.
